Created the database UI
Created the database Ui file. This holds all the elements that make up the database UI section and some of the functionality for them.
